requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
aiohttp==3.9.1
pytest==7.4.3
pytest-mock==3.12.0
responses==0.24.1
mypy==1.7.1 